kubetest: shuffling shared methods around to make further refactors easier
Fix packagename
