+++ kubetest/local.go
@@ -249,4 +249,4 @@ func (n localCluster) GetClusterCreated(gcpProject string) (time.Time, error) {
 	return time.Time{}, errors.New("GetClusterCreated not implemented in localCluster")
 }
 
-func (_ localCluster) KubectlCommand() (*exec.Cmd, error) { return nil, nil }
+func (localCluster) KubectlCommand() (*exec.Cmd, error) { return nil, nil }
